pyproject.toml
[project]
name = "open-review-crawler"
version = "0.1.0"
description = "OpenReview Paper Crawler with MarkItDown"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markitdown[pdf]>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "openreview-py>=1.52.6",
    "pandas>=2.3.3",
    "pydantic>=2.12.0",
]
